Improvements on the folder structure
