+++ app/src/androidTest/java/com/example/apperoSdkAndroid/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.apperoSdkAndroid
+package com.appero.sdk
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.appero_sdk_android", appContext.packageName)
+        assertTrue(appContext.packageName.contains("appero"))
     }
 }
