+++ app/src/main/java/com/appero/sdk/util/DateTimeUtils.kt
@@ -1,4 +1,4 @@
-package com.example.apperoSdkAndroid.utils
+package com.appero.sdk.util
 
 import java.text.SimpleDateFormat
 import java.util.Date
@@ -6,13 +6,9 @@ import java.util.Locale
 import java.util.TimeZone
 
 object DateTimeUtils {
-    /**
-     * Generate current timestamp in ISO 8601 format
-     * @return Formatted timestamp string
-     */
     fun getCurrentTimestamp(): String {
         val dateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.US)
         dateFormat.timeZone = TimeZone.getTimeZone("UTC")
         return dateFormat.format(Date())
     }
-}
+}
